+++ changelog.md
@@ -5,3 +5,10 @@
 - Implemented:
   - Shell of the project.
   - Home route.
+
+## Version 2.0.0
+
+- Implemented:
+  - Connection to our Mongo database.
+  - Models for each collection.
+  - One endpoint to store new careers or get them according to the [documentation](./readme.md).